# Conflicts:
#	Utils/hypothesis_testing.cpp
#	main.cpp

+++ Graphs/UndirectedGraph.h
@@ -59,7 +59,7 @@ private:
     MatrixXd laplacian_matrix;
     MatrixXd sqrt_degree;
     bool diameter_computed = false;
-    int diameter{0};
+    size_t diameter{0};
 
 
 public:
@@ -73,10 +73,10 @@ public:
     MatrixXd get_sqrt_degree_matrix();
     int number_of_nodes();
     int number_of_edges();
-    int estimate_diameter();
+    size_t estimate_diameter();
     map<size_t, string> get_nodes();
     pair<VectorXd, double> get_second_eigenpair();
-    int get_estimated_diameter();
+    size_t get_estimated_diameter();
     set<size_t> sweep_set(VectorXd &second_EV, vector<size_t> degrees);
     pair<UndirectedGraph, UndirectedGraph> cheeger_cut(VectorXd &second_EV);
     void print();
